Add the "list" subcommand
This commit introduces a `list` subcommand that can list either
providers or registered models. The `providers` subcommand lists all
providers currently supported by Crosstalk, along with their enabled
status. The `models` subcommand queries all enabled providers for their
models and generates a list that includes the model ID, provider, and
context length. You can limit the listing to models from a specific
provider using the `--provider` flag.

Additionally, the output can be formatted in three ways: as a table, in
JSON, or as a headerless table. By default, the output is presented in a
'docker-like' table format, which enables filtering and sorting with
standard utilities like `awk`. The header can be omitted for this
purpose. JSON output may be useful for manipulating data with a tool
like `jq` or for processing input in a subprocess.

These features were implemented by querying the registry for available
providers and models, which were then mapped to types in the `list`
module that support serialization and transformation into a CLI table.
This CLI table properly pads the output by adding whitespace. While the
table could be serialized from the `serde` interface, I decided against
it due to the overhead of creating a `serde` serializer, as a CLI table
does not naturally map to Rust objects.