Add Control+E Binding to Launch the Editor
This commit introduces the ability to launch an external editor (e.g.,
`vim` or `emacs`) using the `Control + E` key binding. This feature
supplements the existing `/edit` command. Now, a single temporary file
is created for the entire duration of the session, and the dependency on
the `tempfile` library has been replaced with standard library
primitives.

Unfortunately, two pathways for launching the interactive editor need to
be maintained. `reedline` has an internal method for launching the
editor, which is not publicly exposed. One limitation of this method is
that it writes the current state of the buffer to a temporary file. As a
result, if we wanted to add information about the preceding message to
this file, it would be overwritten.

I explored attaching the `Control + E` binding to an event that would
internally dispatch the `/edit` command and trigger our
`read_from_interactive_editor` method. The issue with this approach is
that there is no way to clear the buffer. When `reedline` issues the
`/edit` command, it doesn't clear the internal buffer. This leads to a
bug in which the previous content of the buffer is printed, even after
the editor exits. While you can issue a "Clear" edit command through
`reedline`, there is no way to refresh the buffer without requiring user
input.